re made a cleaner version of insertion sort
